fix: refactor code to eliminate race condition when fetching mcid
