+++ src/main/kotlin/com/mparticle/kits/AdobeKit.kt
@@ -2,15 +2,36 @@ package com.mparticle.kits
 
 import android.app.Application
 import android.content.Context
-import com.adobe.marketing.mobile.*
+import android.content.Intent
+import com.adobe.marketing.mobile.Analytics
+import com.adobe.marketing.mobile.Identity
+import com.adobe.marketing.mobile.Lifecycle
+import com.adobe.marketing.mobile.Media
 import com.adobe.marketing.mobile.MediaConstants.AdMetadataKeys
 import com.adobe.marketing.mobile.MediaConstants.VideoMetadataKeys
+import com.adobe.marketing.mobile.MediaTracker
+import com.adobe.marketing.mobile.MobileCore
+import com.adobe.marketing.mobile.MobileServices
+import com.adobe.marketing.mobile.Signal
+import com.adobe.marketing.mobile.UserProfile
 import com.mparticle.BaseEvent
 import com.mparticle.MPEvent
-import com.mparticle.media.events.*
+import com.mparticle.MParticle
+import com.mparticle.media.events.ContentType
+import com.mparticle.media.events.EventAttributes
+import com.mparticle.media.events.MediaAd
+import com.mparticle.media.events.MediaAdBreak
+import com.mparticle.media.events.MediaAttributeKeys
+import com.mparticle.media.events.MediaContent
+import com.mparticle.media.events.MediaEvent
+import com.mparticle.media.events.MediaEventName
+import com.mparticle.media.events.MediaSegment
 
-open class AdobeKit: AdobeKitBase(), KitIntegration.EventListener {
+open class AdobeKit: KitIntegration.EventListener, KitIntegration(),
+    KitIntegration.AttributeListener, KitIntegration.PushListener,
+    KitIntegration.ApplicationStateListener {
 
+    internal val MARKETING_CLOUD_ID_KEY = "mid"
     internal val LAUNCH_APP_ID: String = "launchAppId"
 
     var defaultMediaTracker: MediaTracker? = null
@@ -20,7 +41,6 @@ open class AdobeKit: AdobeKitBase(), KitIntegration.EventListener {
     override fun getName() = "Adobe Media"
 
     public override fun onKitCreate(settings: Map<String, String>, context: Context): List<ReportingMessage> {
-        super.onKitCreate(settings, context)
         val appId = settings.get(LAUNCH_APP_ID)
 
         MobileCore.setApplication(context.applicationContext as Application)
@@ -34,8 +54,75 @@ open class AdobeKit: AdobeKitBase(), KitIntegration.EventListener {
         MobileCore.start {
             MobileCore.configureWithAppID(appId)
         }
+        Identity.getExperienceCloudId {
+            setMarketingCloudId(it)
+        }
         defaultMediaTracker = Media.createTracker()
         return listOf()
+    }
+
+    override fun onApplicationForeground() {
+        syncIds()
+    }
+
+    override fun onApplicationBackground() {
+        syncIds()
+    }
+
+    override fun setUserAttribute(s: String, s1: String) {
+        syncIds()
+    }
+
+    override fun setUserAttributeList(s: String, list: List<String>) {
+        syncIds()
+    }
+
+    override fun supportsAttributeLists(): Boolean {
+        return false
+    }
+
+    override fun setAllUserAttributes(map: Map<String, String>, map1: Map<String, List<String>>) {
+        syncIds()
+    }
+
+    override fun removeUserAttribute(s: String) {
+        syncIds()
+    }
+
+    override fun setUserIdentity(identityType: MParticle.IdentityType, s: String) {
+        syncIds()
+    }
+
+    override fun removeUserIdentity(identityType: MParticle.IdentityType) {
+        syncIds()
+    }
+
+    override fun logout(): List<ReportingMessage> = emptyList()
+
+    override fun willHandlePushMessage(intent: Intent): Boolean = false
+
+    override fun onPushMessageReceived(context: Context, intent: Intent) {}
+    override fun onPushRegistration(instanceId: String, senderId: String): Boolean {
+        syncIds()
+        return false
+    }
+
+    private fun syncIds() {
+        Identity.getExperienceCloudId {
+            setMarketingCloudId(it)
+        }
+    }
+
+    override fun getInstance() =
+        Identity.getExperienceCloudId {
+            setMarketingCloudId(it)
+            AdobeApi(it)
+        }
+
+    fun setMarketingCloudId(id: String) {
+        val integrationAttributes = integrationAttributes
+        integrationAttributes[MARKETING_CLOUD_ID_KEY] = id
+        setIntegrationAttributes(integrationAttributes)
     }
 
     override fun setOptOut(optout: Boolean): List<ReportingMessage> = emptyList()
@@ -111,9 +198,9 @@ open class AdobeKit: AdobeKitBase(), KitIntegration.EventListener {
     private fun updateQos(mediaEvent: MediaEvent) {
         mediaEvent.qos?.let { mediaQos ->
             val qoe = Media.createQoEObject(mediaQos.bitRate?.toLong() ?: 0,
-                    mediaQos.startupTime?.toSeconds() ?: 0.0,
-                    mediaQos.fps?.toDouble() ?: 0.0,
-                    mediaQos.droppedFrames?.toLong() ?: 0)
+                mediaQos.startupTime?.toSeconds() ?: 0.0,
+                mediaQos.fps?.toDouble() ?: 0.0,
+                mediaQos.droppedFrames?.toLong() ?: 0)
             mediaTrackers[mediaEvent.sessionId]?.updateQoEObject(qoe)
         }
     }
@@ -174,33 +261,33 @@ open class AdobeKit: AdobeKitBase(), KitIntegration.EventListener {
 
     private fun MediaSegment.getChapterObject(): Map<String?, Any?> {
         return Media.createChapterObject(title,
-                index?.toLong() ?: 0,
-                duration?.toDouble() ?: 0.0,
-                currentPlayheadPosition.toDouble()
+            index?.toLong() ?: 0,
+            duration?.toDouble() ?: 0.0,
+            currentPlayheadPosition.toDouble()
         )
     }
 
     internal fun MediaContent.getMediaObject(): HashMap<String?, Any?> {
         return Media.createMediaObject(
-                name,
-                contentId,
-                duration?.toSeconds() ?: 0.0,
-                streamType,
-                getMediaType()
+            name,
+            contentId,
+            duration?.toSeconds() ?: 0.0,
+            streamType,
+            getMediaType()
         )
     }
 
     internal fun MediaAdBreak.getAdBreakObject(): Map<String?, Any?> {
         return Media.createAdBreakObject(
-                title,
-                1L,
-                currentPlayheadPosition.toSeconds()
+            title,
+            1L,
+            currentPlayheadPosition.toSeconds()
         )
     }
 
     internal fun MediaAd.getAdObject(): Map<String?, Any?> {
         return Media.createAdObject(title, id, position?.toLong() ?: 0, duration?.toDouble()
-                ?: 0.0)
+            ?: 0.0)
     }
 
     internal fun MediaContent.getMediaType(): Media.MediaType? {
@@ -212,30 +299,30 @@ open class AdobeKit: AdobeKitBase(), KitIntegration.EventListener {
     }
 
     internal fun <K: String?, V> Map<K, V>.toAdobeAttributes(): Map<String?, String?> =
-            entries.associate { (key, value) ->
-                when (key) {
-                    MediaAttributeKeys.AD_ADVERTISING_ID -> AdMetadataKeys.ADVERTISER
-                    MediaAttributeKeys.AD_CAMPAIGN -> AdMetadataKeys.CAMPAIGN_ID
-                    MediaAttributeKeys.AD_CREATIVE -> AdMetadataKeys.CREATIVE_ID
-                    MediaAttributeKeys.AD_PLACEMENT -> AdMetadataKeys.PLACEMENT_ID
-                    MediaAttributeKeys.AD_SITE_ID -> AdMetadataKeys.SITE_ID
-                    EventAttributes.CONTENT_SHOW -> VideoMetadataKeys.SHOW
-                    EventAttributes.CONTENT_EPISODE -> VideoMetadataKeys.EPISODE
-                    EventAttributes.CONTENT_ASSET_ID -> VideoMetadataKeys.ASSET_ID
-                    EventAttributes.CONTENT_GENRE -> VideoMetadataKeys.GENRE
-                    EventAttributes.CONTENT_FIRST_AIR_DATE -> VideoMetadataKeys.FIRST_AIR_DATE
-                    EventAttributes.CONTENT_DIGITAL_DATE -> VideoMetadataKeys.FIRST_DIGITAL_DATE
-                    EventAttributes.CONTENT_RATING -> VideoMetadataKeys.RATING
-                    EventAttributes.CONTENT_ORIGINATOR -> VideoMetadataKeys.ORIGINATOR
-                    EventAttributes.CONTENT_NETWORK -> VideoMetadataKeys.NETWORK
-                    EventAttributes.CONTENT_SHOW_TYPE -> VideoMetadataKeys.SHOW_TYPE
-                    EventAttributes.CONTENT_MVPD -> VideoMetadataKeys.MVPD
-                    EventAttributes.CONTENT_AUTHORIZED -> VideoMetadataKeys.AUTHORIZED
-                    EventAttributes.CONTENT_DAYPART -> VideoMetadataKeys.DAY_PART
-                    EventAttributes.CONTENT_FEED -> VideoMetadataKeys.FEED
-                    else -> key
-                } to value.toString()
-            }
+        entries.associate { (key, value) ->
+            when (key) {
+                MediaAttributeKeys.AD_ADVERTISING_ID -> AdMetadataKeys.ADVERTISER
+                MediaAttributeKeys.AD_CAMPAIGN -> AdMetadataKeys.CAMPAIGN_ID
+                MediaAttributeKeys.AD_CREATIVE -> AdMetadataKeys.CREATIVE_ID
+                MediaAttributeKeys.AD_PLACEMENT -> AdMetadataKeys.PLACEMENT_ID
+                MediaAttributeKeys.AD_SITE_ID -> AdMetadataKeys.SITE_ID
+                EventAttributes.CONTENT_SHOW -> VideoMetadataKeys.SHOW
+                EventAttributes.CONTENT_EPISODE -> VideoMetadataKeys.EPISODE
+                EventAttributes.CONTENT_ASSET_ID -> VideoMetadataKeys.ASSET_ID
+                EventAttributes.CONTENT_GENRE -> VideoMetadataKeys.GENRE
+                EventAttributes.CONTENT_FIRST_AIR_DATE -> VideoMetadataKeys.FIRST_AIR_DATE
+                EventAttributes.CONTENT_DIGITAL_DATE -> VideoMetadataKeys.FIRST_DIGITAL_DATE
+                EventAttributes.CONTENT_RATING -> VideoMetadataKeys.RATING
+                EventAttributes.CONTENT_ORIGINATOR -> VideoMetadataKeys.ORIGINATOR
+                EventAttributes.CONTENT_NETWORK -> VideoMetadataKeys.NETWORK
+                EventAttributes.CONTENT_SHOW_TYPE -> VideoMetadataKeys.SHOW_TYPE
+                EventAttributes.CONTENT_MVPD -> VideoMetadataKeys.MVPD
+                EventAttributes.CONTENT_AUTHORIZED -> VideoMetadataKeys.AUTHORIZED
+                EventAttributes.CONTENT_DAYPART -> VideoMetadataKeys.DAY_PART
+                EventAttributes.CONTENT_FEED -> VideoMetadataKeys.FEED
+                else -> key
+            } to value.toString()
+        }
 
     internal fun Long.toSeconds(): Double {
         return toDouble() / 1000
